linkedin_ai_agent/pyproject.toml
[project]
name = "linkedin_ai_agent"
version = "0.1.0"
description = "LinkedIn AI Agent - Multi-Agent Content Creation System with Style Matching"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.11.0,<1.0.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "typing-extensions>=4.5.0",
    "mcp>=1.9.2",
    "python-telegram-bot>=22.1",
    "crewai-tools[mcp]>=0.45.0",
]

[project.scripts]
linkedin_ai_agent = "linkedin_ai_agent.main:run"
run_crew = "linkedin_ai_agent.main:run"
train = "linkedin_ai_agent.main:train"
replay = "linkedin_ai_agent.main:replay"
test = "linkedin_ai_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
